Add post test, change to APITestCase
